isolate edge case test
